move instances from IdentityBoth to AssociativeBoth
