+++ src/main/scala/zio/prelude/IdentityBoth.scala
@@ -66,88 +66,6 @@ object IdentityBoth extends LawfulF.Invariant[DeriveEqualIdentityBothInvariant, 
    */
   def apply[F[_]](implicit identityBoth: IdentityBoth[F]): IdentityBoth[F] =
     identityBoth
-
-  /**
-   * The `IdentityBoth` instance for `Either`.
-   */
-  implicit def EitherIdentityBoth[L]: IdentityBoth[({ type lambda[+r] = Either[L, r] })#lambda] =
-    new IdentityBoth[({ type lambda[+r] = Either[L, r] })#lambda] {
-      val any: Either[L, Any] = Right(())
-
-      def both[A, B](fa: => Either[L, A], fb: => Either[L, B]): Either[L, (A, B)] =
-        fa.flatMap(a => fb.map(b => (a, b)))
-    }
-
-  /**
-   * The `IdentityBoth` instance for a failed `Either`
-   */
-  implicit def EitherFailedIdentityBoth[R]: IdentityBoth[({ type lambda[+l] = Failure[Either[l, R]] })#lambda] =
-    new IdentityBoth[({ type lambda[+l] = Failure[Either[l, R]] })#lambda] {
-      val any: Failure[Either[Any, R]] = Failure.wrap(Left(()))
-
-      def both[A, B](fa: => Failure[Either[A, R]], fb: => Failure[Either[B, R]]): Failure[Either[(A, B), R]] =
-        Failure.wrap {
-          Failure
-            .unwrap(fa)
-            .left
-            .flatMap(a => Failure.unwrap(fb).left.map(b => (a, b)))
-        }
-    }
-
-  /**
-   * The `IdentityBoth` instance for `Future`.
-   */
-  implicit val FutureIdentityBoth: IdentityBoth[Future] =
-    new IdentityBoth[Future] {
-      val any: Future[Any] = Future.successful(())
-
-      def both[A, B](fa: => Future[A], fb: => Future[B]): Future[(A, B)] = fa zip fb
-    }
-
-  /**
-   * The `IdentityBoth` instance for `Option`.
-   */
-  implicit val OptionIdentityBoth: IdentityBoth[Option] =
-    new IdentityBoth[Option] {
-      val any: Option[Any]                                               =
-        Some(())
-      def both[A, B](fa: => Option[A], fb: => Option[B]): Option[(A, B)] =
-        (fa, fb) match {
-          case (Some(a), Some(b)) => Some((a, b))
-          case _                  => None
-        }
-    }
-
-  /**
-   * The `IdentityBoth` instance for `List`.
-   */
-  implicit val ListAssociativeIdentityBoth: IdentityBoth[List] =
-    new IdentityBoth[List] {
-      val any: List[Any]                                           =
-        List(())
-      def both[A, B](fa: => List[A], fb: => List[B]): List[(A, B)] = fa.flatMap(a => fb.map(b => (a, b)))
-    }
-
-  /**
-   * The `IdentityBoth` instance for `Id`.
-   */
-  implicit val IdIdentityBoth: IdentityBoth[Id] =
-    new IdentityBoth[Id] {
-      val any: Id[Any] = Id(())
-
-      def both[A, B](fa: => Id[A], fb: => Id[B]): Id[(A, B)] =
-        Id(Id.unwrap(fa) -> Id.unwrap(fb))
-    }
-
-  /**
-   * The `IdentityBoth` instance for `Try`.
-   */
-  implicit val TryIdentityBoth: IdentityBoth[Try] =
-    new IdentityBoth[Try] {
-      val any: Try[Any] = Success(())
-
-      def both[A, B](fa: => Try[A], fb: => Try[B]): Try[(A, B)] = fa.flatMap(a => fb.map(b => (a, b)))
-    }
 }
 
 trait IdentityBothSyntax {
