+++ src/test/scala/zio/prelude/AssociativeBothSpec.scala
@@ -1,6 +1,5 @@
 package zio.prelude
 
-import zio.prelude.IdentityBoth._
 import zio.test._
 import zio.test.laws._
